+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/membership/controller/ApiV1MembershipController.java
@@ -10,11 +10,12 @@ import org.tuna.zoopzoop.backend.domain.member.dto.etc.SimpleUserInfo;
 import org.tuna.zoopzoop.backend.domain.member.dto.res.ResBodyForGetMemberInfo;
 import org.tuna.zoopzoop.backend.domain.member.entity.Member;
 import org.tuna.zoopzoop.backend.domain.member.service.MemberService;
-import org.tuna.zoopzoop.backend.domain.space.membership.dto.*;
+import org.tuna.zoopzoop.backend.domain.space.membership.dto.etc.SpaceMemberInfo;
+import org.tuna.zoopzoop.backend.domain.space.membership.dto.req.ReqBodyForChangeMemberAuthority;
 import org.tuna.zoopzoop.backend.domain.space.membership.dto.req.ReqBodyForExpelMember;
-import org.tuna.zoopzoop.backend.domain.space.membership.dto.res.ResBodyForExpelMember;
+import org.tuna.zoopzoop.backend.domain.space.membership.dto.req.ReqBodyForInviteMembers;
+import org.tuna.zoopzoop.backend.domain.space.membership.dto.res.*;
 import org.tuna.zoopzoop.backend.domain.space.membership.entity.Membership;
-import org.tuna.zoopzoop.backend.domain.space.membership.enums.Authority;
 import org.tuna.zoopzoop.backend.domain.space.membership.service.MembershipService;
 import org.tuna.zoopzoop.backend.domain.space.space.entity.Space;
 import org.tuna.zoopzoop.backend.domain.space.space.service.SpaceService;
