--- src/main/java/org/tuna/zoopzoop/backend/domain/space/membership/dto/ResBodyForChangeMemberAuthority.java
@@ -1,8 +1,0 @@
-package org.tuna.zoopzoop.backend.domain.space.membership.dto;
-
-public record ResBodyForChangeMemberAuthority(
-        Integer spaceId,
-        String spaceName,
-        SpaceMemberInfo member
-) {
-}
