+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/membership/dto/etc/SpaceMemberInfo.java
@@ -1,4 +1,4 @@
-package org.tuna.zoopzoop.backend.domain.space.membership.dto;
+package org.tuna.zoopzoop.backend.domain.space.membership.dto.etc;
 
 import org.tuna.zoopzoop.backend.domain.space.membership.enums.Authority;
