+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/membership/dto/req/ReqBodyForInviteMembers.java
@@ -1,4 +1,4 @@
-package org.tuna.zoopzoop.backend.domain.space.membership.dto;
+package org.tuna.zoopzoop.backend.domain.space.membership.dto.req;
 
 import java.util.List;
