+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/membership/dto/res/ResBodyForChangeMemberAuthority.java
@@ -1,0 +1,10 @@
+package org.tuna.zoopzoop.backend.domain.space.membership.dto.res;
+
+import org.tuna.zoopzoop.backend.domain.space.membership.dto.etc.SpaceMemberInfo;
+
+public record ResBodyForChangeMemberAuthority(
+        Integer spaceId,
+        String spaceName,
+        SpaceMemberInfo member
+) {
+}
