+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/membership/dto/res/ResBodyForInviteMembers.java
@@ -1,8 +1,6 @@
-package org.tuna.zoopzoop.backend.domain.space.membership.dto;
+package org.tuna.zoopzoop.backend.domain.space.membership.dto.res;
 
 import org.tuna.zoopzoop.backend.domain.member.dto.etc.SimpleUserInfo;
-import org.tuna.zoopzoop.backend.domain.member.dto.res.ResBodyForGetMemberInfo;
-import org.tuna.zoopzoop.backend.domain.space.membership.enums.Authority;
 
 import java.util.List;
