+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/membership/dto/res/ResBodyForSpaceMemberList.java
@@ -1,4 +1,6 @@
-package org.tuna.zoopzoop.backend.domain.space.membership.dto;
+package org.tuna.zoopzoop.backend.domain.space.membership.dto.res;
+
+import org.tuna.zoopzoop.backend.domain.space.membership.dto.etc.SpaceMemberInfo;
 
 import java.util.List;
